refactor(visualizer): move visualizer to new componet at ui lib
