+++ libs/web/visualizer/feature/src/lib/visualizer.component.html
@@ -1,1 +1,1 @@
-<div #visualizer id="visualizer" (dblclick)="toggleFullscreen()"></div>
+<as-web-visualizer-ui></as-web-visualizer-ui>
