+++ libs/web/visualizer/feature/src/lib/visualizer.component.scss
@@ -1,8 +1,3 @@
 :host {
   overflow: hidden;
 }
-
-#visualizer {
-  background: url('/assets/images/noise.png');
-  overflow: hidden;
-}
