+++ libs/web/visualizer/feature/src/lib/visualizer.module.ts
@@ -2,10 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { VisualizerComponent } from './visualizer.component';
 import { RouterModule } from '@angular/router';
+import { WebVisualizerUiModule } from '@angular-spotify/web/visualizer/ui';
 
 @NgModule({
   imports: [
     CommonModule,
+    WebVisualizerUiModule,
     RouterModule.forChild([
       {
         path: '',
