+++ libs/web/visualizer/ui/src/lib/web-visualizer-ui.component.html
@@ -1,0 +1,1 @@
+<div #visualizer id="visualizer" (dblclick)="toggleFullscreen()"></div>
