+++ libs/web/visualizer/ui/src/lib/web-visualizer-ui.component.scss
@@ -1,0 +1,8 @@
+:host {
+  overflow: hidden;
+}
+
+#visualizer {
+  background: url('/assets/images/noise.png');
+  overflow: hidden;
+}
