+++ libs/web/visualizer/ui/src/lib/web-visualizer-ui.module.ts
@@ -1,7 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { WebVisualizerUiComponent } from './web-visualizer-ui.component';
 
 @NgModule({
-  imports: [CommonModule]
+  imports: [CommonModule],
+  declarations: [WebVisualizerUiComponent],
+  exports: [WebVisualizerUiComponent]
 })
 export class WebVisualizerUiModule {}
